+++ RELEASE_NOTES.md
@@ -5,7 +5,7 @@
     * Add Signal Patch Controller
     * Add Signal Patch Requests
     * Add Signal Patch Handlers
-
+    
 #### 1.0 - 16.09.2018
 
   Initial Commit